[XrdTls] Allow disabling of X.509 client auth
The X.509 client auth request is fairly disruptive for communities
that are browser-based (since this is an extremely uncommon flow
for users -- and most of them interpret the certificate request
as an error message) and not using X.509 client auth.

This provides a boolean (defaulting to "on") that allows the administrator
to turn off X.509 client auth when not needed / desired.