Add ResourceMonitor and PurgePlugin
Resource Monitor (main class XrdPfcResourceMonitor) provides detailed and
accurate snapshots of state and usage of existing directories. It allows
optimization of per-directory statistics information collection and storage,
and reduces the number of required cache namespace traversals. It can
periodically export the state into JSON and binary files for viewing of
current state and for further processing or aggregation. This
information also serves as input to purge process.

Purge plugin provides a class-based API that allows users to implement their
own steering of the purge priority selection. The purge is still directory
driven, i.e., the purge plugin is given a per-directory snapshot of resource
monitor state and responds with a list of { directory-path, bytes_to_remove }
structures. Pfc still drives the actual low-level directory traversal and file
purge -- though this could be relaxed in the future. If the plugin does not
schedule sufficient removals, the classic LRU algorithm kicks in to satisfy
the high/low watermark and file-based limits.

The purge-plugin API is defined in class Pfc::PurgePin. Reference
implementation, directory-quota based purging, is provided in the core PFC
library class XrdPfcPurgeQuota and enabled as:
  pfc.purgelib libXrdPfcPurgeQuota /etc/xrootd/quota.cfg
with entries in quota cfg file looking like:
  /store/mc 10T
  /store/data 12T

Co-authored-by: Alja Mrak-Tadel <amraktadel@ucsd.edu>