Export configured FSctl plugin to global configuration
The FSctl plugin is necessary to invoke filesystem-level operations
on the PFC.  However, unlike the `XrdOss` and `XrdAccAuthorize`
plugins, it is only exported in the xrootd protocol environment,
not the global process-wide environment.

This provides symmetry between the FSctl and the other plugins.
It joins other parts of the PFC available via the global export.

I also investigated invoking the plugin indirectly through some
other layer (such as the xrootd protocol bridge) but, unfortunately,
the use case for needing to invoke the plugin is a `XrdHttpExtHandler`;
the handler framework does not provide access to the protocol bridge.