[XrdPfc] Check for a null pointer dereference
When looking up the map of active files, the returned value is a
a `nullptr` if the file-open is ongoing.  Check for this case and
only try to use the file object if it is non-null.

Without this check, the XrdPfc module will trigger a segfault when
trying to increase the reference count for a null object.

The places affected by the nullptr already have appropriate fallback
code in place if the filename isn't on the active list at all -- we
now use that code instead of crashing.