Enable write-through mode for cache
Allow the PFC to proxy ("write-through") writes when configured
to do so (default is current behavior of disabling writes).

The implementation is simple -- when writing is enabled, don't
attach the PFC to the cache object.

Adds a new parameter:
```
pfc.writemode [writethrough | off]
```
where the default write mode is `off`.