Re-engineer concurrency limits for throttles
Previously, the concurrency limits for the throttle was not fair: once they
were hit, all requests were put into the same queue, meaning a single badly-
behaved user could mess up the experience for all users.

Now, we track the per-user recent concurrency and use it to determine how
close to "fair" the load is between all the active users.  If the user has
few active transfers compared to others, they will receive more "shares",
increasing the probability the next randomly-selected user to wake is them.

The concurrency tracking was modernized to remove the use of sequentially
consistent counters (which weren't needed) and switches to acquire/release
semantics instead.  This removes the use of fencing instructions on x86,
meaning the overhead of the throttle tracking is significantly lower than
before under high concurrency.

The list of ongoing operations (used to determine the concurrency) is broken
up across several lists to avoid a single global mutex protecting the operations.

The concurrency itself is calculated using exponential moving averages with
a decay time of 10s.