[#23] Modify XrdThrottle to be an OSS plugin
XrdThrottle was originally designed at the OFS layer because we wanted
to perform loadshedding - redirect a file-open request to another server.

However, the loadshed was rarely used in production and we've hit a
snag: since the authorization checks haven't run, the information about
the user hasn't been extracted from the token.  This means, for token-based
authorization, we map everyone to the session user (typically "nobody").

This change allows the throttle to run as either an OFS or OSS.  When in OSS
mode, we have access to the results of the token authorization and the token subject
information is available for fairshare purposes; redirects are disabled.

If both are loaded, instead of throwing an error, to be nice to existing configuration
files, we throw a warning instead.  The OSS mode can be enabled with the following
line in the configuration file:

```
ofs.osslib ++ libXrdThrottle.so
```

Given the advantages of leaning on the existing OFS, we may consider removing
the OFS layer in the future.