Avoid reference beyond end of table
Previously, when the last entry in the table was deleted, there
was a check of the `N+1` entry.  This could potentially cause a
segfault if the end of the table is the end of a page.

A segfault was triggered when running under ASAN with a high
concurrency of jobs requesting checksums.